// src/api/types.gen.ts
export type BodyTokenAuthTokenPost = {
    /**
     * Refresh Token
     */
    refresh_token?: string | null;
    /**
     * Grant Type
     */
    grant_type: string;
    /**
     * Code
     */
    code?: string | null;
    /**
     * Redirect Uri
     */
    redirect_uri?: string | null;
    /**
     * Client Id
     */
    client_id?: string | null;
    /**
     * Client Secret
     */
    client_secret?: string | null;
    /**
     * Code Verifier
     */
    code_verifier?: string | null;
};
// ...
export type TokenResponse = {
    /**
     * Access Token
     */
    access_token: string;
    /**
     * Token Type
     */
    token_type?: 'bearer';
    /**
     * Expires In
     */
    expires_in?: number;
    /**
     * Scope
     */
    scope?: string;
    /**
     * Refresh Token
     */
    refresh_token: string;
    /**
     * Id Token
     */
    id_token?: string | null;
};